ARM: plat: fix build warning in reg_rw.c
Fix below build warning.
arch/arm/plat-pxa/reg_rw.c: In function 'pxa_reg_add':
arch/arm/plat-pxa/reg_rw.c:64: warning: assignment makes integer from pointer without a cast

+++ arch/arm/plat-pxa/reg_rw.c
@@ -61,7 +61,7 @@ int pxa_reg_add(u32 pa, u32 mask)
 	r->pa = pa;
 	r->wo_mask = mask;
 	r->written = 0xffffffff & r->wo_mask;
-	r->kva = ioremap(pa,sizeof(u32));
+	r->kva = (u32) ioremap(pa,sizeof(u32));
 	r->val = readl(r->kva);
 
 	list_add(&r->list, &reg_list);
